Add CMakeLists.txt for building the project
This commit adds the CMakeLists.txt file to the project, which is used to build the project using CMake. The file sets the minimum required CMake version, specifies the project name, sets the C++ standard to C++11, finds the OpenCV package, collects all source and header files in the 'src' directory, adds an executable target, and links the OpenCV libraries to the target.

With this file, users can build the project using CMake by running the following commands in the project root directory:

mkdir build
cd build
cmake ..
cmake --build .

This will create a 'build' directory, configure the project using CMake, and build the executable target.