Create script/ with placeholder files for future use
This commit adds an empty folder with placeholder files to the Billiard Computer Vision Sports Analyzer project. This folder includes:
- scripts: for any build scripts or utility scripts that may be created

By adding this folder now with placeholder files, we ensure that the complete folder structure is in the repository, even if some folders are currently empty. This will make it easier to add appropriate files to this folder as the project develops further.