Create scripts/ with placeholder files for future use
This commit adds empty folders with placeholder files to the Billiard Computer Vision Sports Analyzer project. These folders include:

- include: for header files that are not in "src"
- lib: for any third-party libraries that may be needed in the future
- test: for unit tests or integration tests
- docs: for additional documentation files
- scripts: for any build scripts or utility scripts that may be created

By adding these folders now with placeholder files, we ensure that the complete folder structure is in the repository, even if some folders are currently empty. This will make it easier to add appropriate files to these folders as the project develops further.